[DO NOT MERGE] example of splitting code for tree shaking
